+++ embassy-stm32/CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## Unreleased - ReleaseDate
 
-- fix: Fixed STM32H5 builds requiring time feature
+- feat: Add blocking read/write methods for STM32 CAN (bxCAN and FDCAN)
 
 ## 0.4.0 - 2025-08-26